requirements.txt
ollama
simpleaudio
piper-tts
pygame
TTS
